+++ options/index.js
@@ -195,8 +195,6 @@ async function init() {
 
   const $defaultLoaded = document.querySelector('#defaults-loaded-msg');
 
-  let $soundTip = document.querySelector('.sound-tips');
-
   function hideFloatMenus() {
     events.toggleOptionMenu.bind(events)(null, false);
     permissions.toggleDialog.bind(permissions)($permsBtn, false);
@@ -260,11 +258,6 @@ async function init() {
   events.addObserver('load', permissions.update.bind(permissions));
   events.addObserver('requestPermissions', permissions.request.bind(permissions));
   permissions.addObserver('update', events.updateAvailability.bind(events))
-
-  $soundTip.querySelector('span[data-i18n$="tips_ctrlPlay"]').innerHTML = browser.i18n.getMessage('options_sound_tips_ctrlPlay');
-  $soundTip.querySelector('button.dismiss').addEventListener('click', e => {
-    $soundTip.classList.add('hidden');
-  });
 
   $save.addEventListener('click', () => {
     $save.disabled = true;